XamlUICommand and related
