+++ FluentAvalonia/UI/Input/CanExecuteRequestedEventArgs.cs
@@ -2,6 +2,9 @@
 
 namespace FluentAvalonia.UI.Input
 {
+	/// <summary>
+	/// Provides event data for the CanExecuteRequested event.
+	/// </summary>
 	public class CanExecuteRequestedEventArgs : EventArgs
 	{
 		internal CanExecuteRequestedEventArgs(object param)
@@ -9,7 +12,14 @@ namespace FluentAvalonia.UI.Input
 			Parameter = param;
 		}
 
+		/// <summary>
+		/// Gets or sets a value indicating whether the ICommand that raised this event is able to execute.
+		/// </summary>
 		public bool CanExecute { get; set; } = true;
+
+		/// <summary>
+		/// Gets the command parameter passed into the CanExecute method that raised this event.
+		/// </summary>
 		public object Parameter { get; }
 	}
 }
