+++ FluentAvalonia/UI/Input/ExecuteRequestedEventArgs.cs
@@ -2,6 +2,9 @@
 
 namespace FluentAvalonia.UI.Input
 {
+	/// <summary>
+	/// Provides event data for the ExecuteRequested event.
+	/// </summary>
 	public class ExecuteRequestedEventArgs : EventArgs
 	{
 		internal ExecuteRequestedEventArgs(object param)
@@ -9,6 +12,9 @@ namespace FluentAvalonia.UI.Input
 			Parameter = param;
 		}
 
+		/// <summary>
+		/// Gets the command parameter passed into the Execute method that raised this event.
+		/// </summary>
 		public object Parameter { get; }
 	}
 }
